corrected eslint errors
