+++ src/api/sshKeys.ts
@@ -1,14 +1,7 @@
 import express from 'express'
-import fs from 'fs'
-import os from 'os'
 
-import {
-  authorizedKeys,
-  addAuthorizedKey,
-  setAuthorizedKeys
-} from '../helpers/authorizedKeys'
+import { authorizedKeys, addAuthorizedKey } from '../helpers/authorizedKeys'
 
-const userHomeDirectory = os.homedir()
 const sshKeyRouter = express.Router()
 
 // The steps below are covered by the setup script. This is not necessssary.
@@ -20,22 +13,6 @@ const sshKeyRouter = express.Router()
 
 sshKeyRouter.get('/', (req, res) => {
   // Read from file if authorizedKeys is empty
-  if (!authorizedKeys) {
-    fs.readFile(`${userHomeDirectory}/.ssh/authorized_keys`, (error, data) => {
-      if (error) {
-        console.log(error)
-      }
-      const keysObj = {}
-      data
-        .toString()
-        .split('\n')
-        .filter(e => e !== '')
-        .forEach((item, index) => {
-          keysObj[`default+${index}`] = item
-        })
-      setAuthorizedKeys(keysObj)
-    })
-  }
   res.json(authorizedKeys)
 })
